+++ spec/Pim/Bundle/CatalogBundle/MongoDB/Normalizer/ProductValueNormalizerSpec.php
@@ -18,8 +18,10 @@ class ProductValueNormalizerSpec extends ObjectBehavior
         $this->shouldImplement('Symfony\Component\Serializer\Normalizer\NormalizerInterface');
     }
 
-    function it_supports_normalization_in_mongodb_json_of_value(ProductValueInterface $value)
+    function it_supports_normalization_in_mongodb_json_of_value(ProductValueInterface $value, AbstractAttribute $attribute)
     {
+        $attribute->getBackendType()->willReturn('foo');
+
         $this->supportsNormalization($value, 'mongodb_json')->shouldBe(true);
         $this->supportsNormalization($value, 'json')->shouldBe(false);
         $this->supportsNormalization($value, 'xml')->shouldBe(false);
@@ -34,6 +36,7 @@ class ProductValueNormalizerSpec extends ObjectBehavior
         $this->setSerializer($serializer);
 
         $attribute->getCode()->willReturn('code');
+        $attribute->getBackendType()->willReturn('foo');
         $attribute->isLocalizable()->willReturn(false);
         $attribute->isScopable()->willReturn(false);
 
@@ -86,4 +89,21 @@ class ProductValueNormalizerSpec extends ObjectBehavior
 
         $this->normalize($value, 'mongodb_json', [])->shouldReturn(null);
     }
+
+    function it_normalizes_value_with_decimal_support_backend(
+        ProductValueInterface $value,
+        AbstractAttribute $attribute
+    ) {
+        $attribute->getCode()->willReturn('code');
+        $attribute->getBackendType()->willReturn('decimal');
+        $attribute->isLocalizable()->willReturn(false);
+        $attribute->isScopable()->willReturn(false);
+
+        $value->getData()->willReturn('42.42');
+        $value->getAttribute()->willReturn($attribute);
+
+        $this->normalize($value, 'mongodb_json', [])->shouldReturn(['code' => 42.42]);
+    }
+
+
 }